react-quiz reset
