+++ src/App.js
@@ -1,0 +1,8 @@
+import DateCounter from "./DateCounter";
+export default function App() {
+  return (
+    <div>
+      <DateCounter />
+    </div>
+  );
+}
