--- src/components/Footer.js
@@ -1,9 +1,0 @@
-function Footer({ children }) {
-  return (
-    <div>
-      <footer>{children}</footer>
-    </div>
-  );
-}
-
-export default Footer;
