--- src/components/Main.js
@@ -1,5 +1,0 @@
-function Main({ children }) {
-  return <main className="main">{children}</main>;
-}
-
-export default Main;
