--- src/components/Question.js
@@ -1,12 +1,0 @@
-import Options from "./Options";
-
-function Question({ question, dispatch, answer }) {
-  return (
-    <div>
-      <h4>{question.question}</h4>
-      <Options question={question} dispatch={dispatch} answer={answer} />
-    </div>
-  );
-}
-
-export default Question;
